Add explicit param for key auth + no overwrite
Key auth can now be explicitely disabled, rather than implicitly
disabled by passing an empty key (which is reasonably brittle).
Exceptions are raised when we end up in a situation we can't handle
(e.g. key auth is enabled but no key is provided).

Existing keys are now preserved, this makes it possible to prepare the
key ahead of time.